+++ src/index.ts
@@ -4,6 +4,7 @@
 //  NOTICE: You are permitted to use, modify, and distribute this file
 //  in accordance with the terms of the license agreement accompanying it.
 // ------------------------------------------------------------------------------
+
 /**
  * PalidorPixiExtension Extension
  */
@@ -13,6 +14,6 @@ export { IFlowViewMapping } from "./robotlegs/bender/extensions/palidorPixi/api/
 export { PalidorEvent } from "./robotlegs/bender/extensions/palidorPixi/events/PalidorEvent";
 export { FlowManager } from "./robotlegs/bender/extensions/palidorPixi/impl/FlowManager";
 export { FlowViewMapping } from "./robotlegs/bender/extensions/palidorPixi/impl/FlowViewMapping";
-export { PixiContainerContoller } from "./robotlegs/bender/extensions/palidorPixi/impl/PixiContainerContoller";
+export { PixiContainerController } from "./robotlegs/bender/extensions/palidorPixi/impl/PixiContainerController";
 export { PalidorPixiExtension } from "./robotlegs/bender/extensions/palidorPixi/PalidorPixiExtension";
 export { PalidorBundle } from "./robotlegs/bender/bundles/palidor/PalidorBundle";